server/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
python-socketio==5.9.0
python-engineio==4.7.1
python-dotenv==1.0.0
Werkzeug==2.3.7
psycopg2-binary==2.9.7
